refactor(react-server): rename features/client-component
