+++ examples/react-server/src/basic.test.tsx
@@ -3,7 +3,7 @@ import { Window } from "happy-dom";
 import React from "react";
 import reactDomClient from "react-dom/client";
 import { beforeAll, beforeEach, expect, test } from "vitest";
-import { initializeWebpackBrowser } from "./features/use-client/browser";
+import { initializeWebpackBrowser } from "./features/client-component/browser";
 
 beforeAll(() => {
   // doing the same as examples/react-server/src/features/bootstrap/plugin.ts
