+++ examples/react-server/src/entry-browser.tsx
@@ -2,9 +2,9 @@ import "virtual:unocss.css";
 import React from "react";
 import reactDomClient from "react-dom/client";
 import type { StreamData } from "./entry-server";
+import { initializeWebpackBrowser } from "./features/client-component/browser";
 import { listenWindowHistory } from "./features/router/browser";
 import { RouterContext } from "./features/router/client";
-import { initializeWebpackBrowser } from "./features/use-client/browser";
 import { readStreamScript } from "./features/utils/stream-script";
 import { $__global } from "./global";
