+++ examples/react-server/src/features/client-component/browser.ts
