+++ examples/react-server/src/features/client-component/server.ts
